requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0